Moves callbacks for events outside main.js (#105)
This could potentially break something with paths. I recall that being an issue, with builds, but just making a note here for reference to future selves when things aren't working. With not legitimate testing in place the only way to test if this works is to do a build and deploy a cluster via the binary executable.